Move ContractCaller tests to their own file
